+++ src/main/java/carracing/domain/carracing/Position.java
@@ -1,4 +1,4 @@
-package carracing.car;
+package carracing.domain.carracing;
 
 import java.util.Objects;
