+++ src/main/java/carracing/view/ResultView.java
@@ -1,8 +1,9 @@
-package carracing;
+package carracing.view;
 
-import carracing.car.Car;
-import carracing.car.Cars;
-import carracing.car.Position;
+import carracing.domain.carracing.CarsMoveStatusHistory;
+import carracing.domain.carracing.Car;
+import carracing.domain.carracing.Cars;
+import carracing.domain.carracing.Position;
 
 import java.util.List;
 import java.util.StringJoiner;
