+++ src/test/java/carracing/domain/carracing/CarRacingTest.java
@@ -1,4 +1,4 @@
-package carracing;
+package carracing.domain.carracing;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
