+++ src/test/java/carracing/domain/carracing/NameTest.java
@@ -1,6 +1,5 @@
-package carracing.car;
+package carracing.domain.carracing;
 
-import carracing.car.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
